14. usunięcie klasy Node.java oraz wpisanie ją do klasy IntLinedList.java
