+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -49,4 +49,31 @@ public class IntLinkedList {
         return i;
     }
 
+     static class Node {
+
+        private int value;
+        private Node prev;
+        public Node next;
+
+        public Node(int i) {
+            setValue(i);
+        }
+
+        public int getValue() {
+            return value;
+        }
+
+        public void setValue(int value) {
+            this.value = value;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+    }
+
 }
