--- src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -1,28 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public class Node {
-
-    private final int value;
-    private Node prev;
-    public Node next;
-
-    public Node(int i) {
-        setValue(i);
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    public void setValue(int value) {
-        this.value = value;
-    }
-
-    public Node getPrev() {
-        return prev;
-    }
-
-    public void setPrev(Node prev) {
-        this.prev = prev;
-    }
-}
